Created initial PetListTable
